[Analysis] Use int64 instead of uint64 for GitilesCommit.position.
uint64 is not compatible with the BigQuery Integer type. The
additional range afforded by this type over int64 is not really
important, so drop it back to an int64 and use validation to
prevent use of negative values.

Proto is quite new and currently only used in pRPC transfers, so
replacement should be safe.

BUG=b:279110567
TEST=ResultDB + LUCI Analysis Integration tests.

Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4501680
Commit-Queue: Beining Chen <beining@google.com>
Auto-Submit: Patrick Meiring <meiring@google.com>

+++ resultdb/internal/services/testmetadataupdator/updator.go
@@ -232,7 +232,7 @@ func (u *testMetadataUpdator) testMetadataMutation(project, testID, subRealm str
 		spanner.CommitTimestamp,
 		spanutil.Compressed(pbutil.MustMarshal(tm)).ToSpanner(),
 		spanutil.Compressed(pbutil.MustMarshal(u.sourceRef())).ToSpanner(),
-		int64(u.sources.GitilesCommit.Position),
+		u.sources.GitilesCommit.Position,
 	}
 	return spanner.InsertOrUpdate("TestMetadata", saveCols, vals)
 }
